+++ src/main/java/com/buuz135/industrial/gui/transporter/GuiTransporter.java
@@ -31,7 +31,6 @@ import com.buuz135.industrial.proxy.block.filter.IFilter;
 import com.buuz135.industrial.proxy.network.TransporterButtonInteractMessage;
 import com.buuz135.industrial.utils.Reference;
 import com.mojang.blaze3d.systems.RenderSystem;
-import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
@@ -85,7 +84,7 @@ public class GuiTransporter extends AbstractContainerScreen<ContainerTransporter
         guiGraphics.blit(BG_TEXTURE, x, y, 0, 0, imageWidth, imageHeight);
         if (upgrade != null) {
             String localized = Component.translatable(upgrade.getFactory().getUpgradeItem().getDescriptionId()).getString();
-            guiGraphics.drawString(Minecraft.getInstance().font, localized, x + imageWidth / 2 - getMinecraft().font.width(localized) / 2, y + 6, 0x404040);
+            guiGraphics.drawString(Minecraft.getInstance().font, localized, x + imageWidth / 2 - getMinecraft().font.width(localized) / 2, y + 6, 0x404040, false);
         }
         for (IGuiComponent iGuiComponent : componentList) {
             iGuiComponent.drawGuiBackgroundLayer(guiGraphics, x, y, mouseX, mouseY);