+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# Version 3.5.16
+
+* Fixed Shiny versions of tools not being added to Infinity Tools
+* Changed Enchantment Applicator slot to limit to 1, closes #1419
+* Fixed end version of ores in the Laser Ore Drill having an extra tag, closes #1417
+* Changed Potion Brewer slot to limit to 1 so bottle don't stack when a mod changes that, closes #1413
+* Added support for mangrove trees in the plant gatherer, closes #1412
+* Fixed Cast Exception on player renderer, closes #1407
+* Added a tag cache for ore fluid, closes #1405
+
 # Version 3.5.13
 
 * Fixed Plant Gatherer skipping blocks